Go: Add Mx() for Noms Maps (#1876)
As one step towards #1819, we've created MapMutator, which can take a
bunch of (what would normally be) Map.Set() calls and batch them up to
be applied all at once. The keys and values are held in a LevelDB cache
until everything's done. Usage looks like this:

 m := types.NewMap()
 mx := m.Mx()
 mx = mx.Set(String("foo"), String("bar")).Set(String("baz"), Number(42))
 m = mx.Finish()

We intend to make this the only way to modify collections, but at first
this will only work on an empty NomsMap.

+++ go/types/number.go
@@ -17,7 +17,7 @@ func (v Number) Equals(other Value) bool {
 
 func (v Number) Less(other Value) bool {
 	if v2, ok := other.(Number); ok {
-		return float64(v) < float64(v2)
+		return v < v2
 	}
 	return NumberKind < other.Type().Kind()
 }
